backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
geographiclib==2.0
geopy==2.4.1
groq==0.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tomlkit==0.11.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==3.0.2
youtube-transcript-api==0.6.2
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
executing==2.0.1
filelock==3.13.1
fireworks-ai==0.14.0
Flask==3.0.2
Flask-Cors==4.0.0
geographiclib==2.0
geopy==2.4.1
groq==0.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
idna==3.6
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.23.2
packaging==24.0
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pymongo==4.6.3
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tomlkit==0.11.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==3.0.2
youtube-transcript-api==0.6.2
